Implementation of user attribute on SmearSlide domain class.
Added create() and list() methods to SmearSlideController to expose only
that user's slides in list view, or all slides if user has ROLE_ADMIN,
and to assign a new slide to the logged-in user.

Instantiated an explicit (not scaffolded) list.gsp for SmearSlide. 

Modifications to _form.gsp template used in create.gsp and update.gsp to
support new user-related functionality.